+++ NEWS.md
@@ -9,6 +9,7 @@
 - Attribute specifications for `all` are now parsed correctly (#889).
 - Predefined `"="` and `"/="` operators are no longer declared for file
   types.
+- Alias of packages now work correctly (#893).
 
 ## Version 1.12.2 - 2024-05-15
 - Fixed a crash when `'transaction` is used with a record type.